feat: refractor home screen, move files
